requirements.txt
google-adk>=1.8.0
google-cloud-dataproc>=5.7.0
google-cloud-bigquery>=3.25.0
google-cloud-logging>=3.10.0
google-cloud-monitoring>=2.20.0
google-cloud-storage>=2.17.0
python-dateutil>=2.9.0

